Added authentication center and cleaned up the code

+++ node4/node.py
@@ -17,10 +17,8 @@ except ImportError:
 class Node:
 
     #--------GLOBAL VARIABLES----------------
-
-    #TCP_IP = '127.0.0.1'  # my IP 164.15.244.54
-    TCP_PORT = 5005        # port number used for the TCP authentication center connection
-    TCP_PORT_BROAD = 5004  # testing for broadcoast 
+    
+    TCP_PORT = 5003
     BUFFER_SIZE = 1024     # size of the receiveng buffer -- we can adapt it to the lenght
                            # of our messages witch will speed up the transition
     Password = b'Dricot'   # users password
@@ -180,7 +178,6 @@ class Node:
                 nodesMessage.start()
                 nodesMessage.join()
                 self.confirmed.clear()
-
         else:
             if receivedBlock in self.blockchain.waiting_blocks:
                 if addr not in self.confirmed:
@@ -231,41 +228,6 @@ class Node:
                 break
         return verified
             
-            
-        
-
-    
-    def runAuthenticationCenterCom(self):
-        """ fonction run as a thread it's responsible for the communication with
-        the authentication center
-        """
-        
-        while True:
-            answer = input("Write 't' for transaction or 'l' to logout")
-            if answer == 't':
-                authen = self.authenticate()
-                if authen == b'ok':
-                    blockchain = Blockchain()
-                    value = input("Write the value of the transaction: ")
-                    blockchain.new_transaction(value)
-            elif answer == 'l':
-                break
-            else:
-                print('Wrong command')
-                pass
-
-    def usersAction(answer):
-        if answer == 't':
-            authen = self.authenticate()
-            if authen == b'ok':
-                blockchain = Blockchain()
-                value = input("Write the value of the transaction: ")
-                blockchain.new_transaction(value)
-##        elif answer == 'l':
-##            
-##        else:
-##            print('Wrong command')
-##            pass
 
 
     def authenticate(self):
@@ -275,29 +237,44 @@ class Node:
 
         
         s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        s.connect((self.server_address, self.TCP_PORT))
-        message = str({'Username':self.username,'Request': 'send nonce motherfucker'}).encode('utf-8') # User's request of nonce
-        s.send(message)
-        data = s.recv(self.BUFFER_SIZE)
-        hashedMessage = hashlib.sha256()
-        hashedMessage.update(data)
-        hashedMessage.update(self.Password)  # The received nonce and the password are concatenated and hashed
-        myHash = hashedMessage.digest()  #or hexdigest for a more condensed form
-        s.send(myHash)
-        data = s.recv(self.BUFFER_SIZE)
-        print ("received authentication:", data) # Response of the authentication center
-        s.close()
-        return data
+        while True:
+            try:
+                s.connect((self.server_address, self.TCP_PORT))
+                message = str({'Username':self.username,'Request': 'send nonce'}).encode('utf-8') # User's request of nonce
+                s.send(message)
+                data = s.recv(self.BUFFER_SIZE)
+                hashedMessage = hashlib.sha256()
+                hashedMessage.update(data)
+                hashedMessage.update(self.Password)  # The received nonce and the password are concatenated and hashed
+                myHash = hashedMessage.digest()  #or hexdigest for a more condensed form
+                s.send(myHash)
+                data = s.recv(self.BUFFER_SIZE)  # Response of the authentication center
+                s.close()
+                return data
+                break
+            except TimeoutError:
+                pass
+            except ConnectionRefusedError:
+                pass
+        socketNodes.close()
+
 
 
    
 
     def sendMessage(self, message):
         s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        print(self.server_address, self.TCP_PORT)
-        s.connect((self.server_address, self.TCP_PORT))
-        s.send(message)
-        s.close()
+        while True:
+            try:
+                s.connect((self.server_address, self.TCP_PORT))
+                s.send(message)
+                s.close()
+                break
+            except TimeoutError:
+                pass
+            except ConnectionRefusedError:
+                print('connection refused')
+                pass
 
     def runInterface(self):
         tran_op = 0
@@ -307,52 +284,45 @@ class Node:
                 print("Do you want to make a transaction?")
                 transac_status = input("")
                 if transac_status =="yes":
-                    #athentication = node.authenticate()
-                    #if authenticate == b'ok'
-                    print("How much money do you want to transfer?")
-                    amount = input("")
-                    if amount.isdigit():                
-                        if amount == 0:
-                            tran_op = 1
-                            print("Transaction Impossible - Amount Null")
-                            print(self.blockchain.last_block)
-                            
-                        else:                                       
-                            
-                            """
-                            Here we create a new block that we broadcoast.
-                            """
-                            
-                            
-                            if (self.blockchain.chain == []):                  # ïf blockchain is empty, create genesis block
-                                #blockchain.create_genesis_block()
-                                nodesMessage = Thread(target = self.runNodesMessage) #Problem. We kill the last thread even if it didn't accomplished the task
-                                self.message = self.setMessage((self.ip_address,{'Block':self.blockchain.create_genesis_block(amount)}))
-                                nodesMessage.setDaemon(True)
-                                nodesMessage.start()
-                            else:
+                    authentication = self.authenticate()
+                    if authentication == b'ok':
+                        print("How much money do you want to transfer?")
+                        amount = input("")
+                        if amount.isdigit():                
+                            if amount == 0:
+                                tran_op = 1
+                                print("Transaction Impossible - Amount Null")
+                                print(self.blockchain.last_block)
+                                
+                            else:                                       
+                                
+                                """
+                                Here we create a new block that we broadcoast.
+                                """
+                                
+                                
+                                if (self.blockchain.chain == []):                  # ïf blockchain is empty, create genesis block
+                                    nodesMessage = Thread(target = self.runNodesMessage) #Problem. We kill the last thread even if it didn't accomplished the task
+                                    self.message = self.setMessage((self.ip_address,{'Block':self.blockchain.create_genesis_block(amount)}))
+                                    nodesMessage.setDaemon(True)
+                                    nodesMessage.start()
+                                else:
 
-                                #proof = node.blockchain.proof_of_work(node.blockchain.last_block())[0]
-                                #time_proof = node.blockchain.proof_of_work(node.blockchain.last_block())[1]
-                                #print(self.blockchain.chain)
-                                proof, time_proof = self.blockchain.proof_of_work()
-                                previous_hash = self.blockchain.last_block['hash']
-                                #blockchain.new_block(blockchain.last_block['hash'])
-                                newBlock = self.blockchain.new_block(amount, proof, time_proof, previous_hash)
-                                nodesMessage = Thread(target = self.runNodesMessage) #Problem. We kill the last thread even if it didn't accomplished the task
-                                self.message = self.setMessage((self.ip_address,{'Block':newBlock}))
-                                nodesMessage.setDaemon(True)
-                                nodesMessage.start()
-                            print("Transaction Validated")   
-    ##                        timer.setTime(0)
-    ##                        while (timer.getTime() < 30):
-    ##                            if False:                                 # insert block incoming condition here
-    ##                                blockchain.chain[-1] = []             # deletes created block if it recieves another block
-    ##                        """append block"""
-                        
+                                    proof, time_proof = self.blockchain.proof_of_work()
+                                    previous_hash = self.blockchain.last_block['hash']
+                                    newBlock = self.blockchain.new_block(amount, proof, time_proof, previous_hash)
+                                    nodesMessage = Thread(target = self.runNodesMessage) #Problem. We kill the last thread even if it didn't accomplished the task
+                                    self.message = self.setMessage((self.ip_address,{'Block':newBlock}))
+                                    nodesMessage.setDaemon(True)
+                                    nodesMessage.start()
+                                print("Transaction Validated")   
+                                
+                        else:
+                            tran_op = 1;
+                            print("Transaction Impossible - Wrong Input")
+
                     else:
-                        tran_op = 1;
-                        print("Transaction Impossible - Wrong Input")
+                        print("Non authenticated user.")
                 elif transac_status =="no":
                     print("Ending Transaction")
                     tran_op = 1
@@ -388,29 +358,12 @@ class Node:
 def main():
 
     node = Node()   
-    #MESSAGE = str({'Username':node.username,'Password':node.Password}).encode('utf-8')
-    #node.sendMessage(MESSAGE)
-
-    #blockchain = Blockchain()
-
-    #authenticationCenterCom = Thread(target = self.runAuthenticationCenterCom)
+    MESSAGE = str({'Username':node.username,'Password':node.Password}).encode('utf-8')
+    node.sendMessage(MESSAGE)
     
     nodeListener = Thread(target = node.runNodesListener)        
-    #nodesMessage = Thread(target = node.runNodesMessage)
-    #authenticationCenterCom.setDaemon(True)
-    #nodeListener.setDaemon(True)
-    #nodesMessage.setDaemon(True)
-    #authenticationCenterCom.start()
     
-    
-    nodeListener.start()
-    #nodesMessage.start()
-    #print('ok')
-    #timer.start()
-    
-
-
-    
+    nodeListener.start()    
             
     interface = Thread(target = node.runInterface)
     interface.start()
